# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ZKAttendance API",
    "description": "Configuration, device and attendance endpoints for the biometric attendance system.",
    "version": "v1"
  },
  "paths": {
...
    "/api/System/devices/{deviceId}/test": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Ping and port-test one device right now, without waiting for the timer.",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/System/sync/{deviceId}": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Run a sync round against a single device.",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/System/nepali-date": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Convert a Gregorian date to Bikram Sambat.",
        "description": "Supported range is 2080-2086 BS. Outside it the converter throws,\r\ndeliberately, rather than returning a wrong date.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Gregorian date, yyyy-MM-dd. Defaults to today.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/System/nepali-date/month-range": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "First and last Gregorian date of a Bikram Sambat month.",
        "description": "Used by monthly reports, which must run Baisakh 1 to the\r\n            end of Baisakh rather than the 1st to the 31st of an AD month.",
        "parameters": [
          {
            "name": "bsYear",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "bsMonth",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {

        }
      }
    }
  }
}